Corrected type return

+++ cpp_src/mod.cpp
@@ -12,9 +12,9 @@
 #include "src/day02.hpp"
 
 extern "C" {
-    // Day01 part 2
+    // Day02 part 2
     EXPORT_FUNC
-    uint32_t count_repeated_digits(const char* payload_ptr, int payload_len) {
+    uint64_t count_repeated_digits(const char* payload_ptr, int payload_len) {
         std::string_view payload(payload_ptr, payload_len);
         return day02::day02_pt1(payload);
     }
